go1 release

+++ misc/dashboard/builder/main.go
@@ -5,14 +5,13 @@
 package main
 
 import (
+	"bytes"
 	"encoding/xml"
-	"errors"
 	"flag"
 	"fmt"
 	"io/ioutil"
 	"log"
 	"os"
-	"path"
 	"path/filepath"
 	"regexp"
 	"runtime"
@@ -24,7 +23,7 @@ import (
 const (
 	codeProject      = "go"
 	codePyScript     = "misc/dashboard/googlecode_upload.py"
-	hgUrl            = "https://go.googlecode.com/hg/"
+	hgUrl            = "https://code.google.com/p/go/"
 	mkdirPerm        = 0750
 	waitInterval     = 30 * time.Second // time to wait before checking for new revs
 	pkgBuildInterval = 24 * time.Hour   // rebuild packages every 24 hours
@@ -33,30 +32,27 @@ const (
 // These variables are copied from the gobuilder's environment
 // to the envv of its subprocesses.
 var extraEnv = []string{
-	"GOHOSTOS",
-	"GOHOSTARCH",
-	"PATH",
-	"DISABLE_NET_TESTS",
-	"MAKEFLAGS",
 	"GOARM",
+	"GOHOSTARCH",
+	"GOHOSTOS",
+	"PATH",
+	"TMPDIR",
 }
 
 type Builder struct {
 	name         string
 	goos, goarch string
 	key          string
-	codeUsername string
-	codePassword string
 }
 
 var (
-	buildroot     = flag.String("buildroot", path.Join(os.TempDir(), "gobuilder"), "Directory under which to build")
+	buildroot     = flag.String("buildroot", filepath.Join(os.TempDir(), "gobuilder"), "Directory under which to build")
 	commitFlag    = flag.Bool("commit", false, "upload information about new commits")
 	dashboard     = flag.String("dashboard", "build.golang.org", "Go Dashboard Host")
 	buildRelease  = flag.Bool("release", false, "Build and upload binary release archives")
 	buildRevision = flag.String("rev", "", "Build specified revision and exit")
-	buildCmd      = flag.String("cmd", "./all.bash", "Build command (specify absolute or relative to go/src/)")
-	external      = flag.Bool("external", false, "Build external packages")
+	buildCmd      = flag.String("cmd", filepath.Join(".", allCmd), "Build command (specify relative to go/src/)")
+	failAll       = flag.Bool("fail", false, "fail all builds")
 	parallel      = flag.Bool("parallel", false, "Build multiple targets in parallel")
 	verbose       = flag.Bool("v", false, "verbose")
 )
@@ -65,6 +61,9 @@ var (
 	goroot      string
 	binaryTagRe = regexp.MustCompile(`^(release\.r|weekly\.)[0-9\-.]+`)
 	releaseRe   = regexp.MustCompile(`^release\.r[0-9\-.]+`)
+	allCmd      = "all" + suffix
+	cleanCmd    = "clean" + suffix
+	suffix      = defaultSuffix()
 )
 
 func main() {
@@ -77,7 +76,7 @@ func main() {
 	if len(flag.Args()) == 0 && !*commitFlag {
 		flag.Usage()
 	}
-	goroot = path.Join(*buildroot, "goroot")
+	goroot = filepath.Join(*buildroot, "goroot")
 	builders := make([]*Builder, len(flag.Args()))
 	for i, builder := range flag.Args() {
 		b, err := NewBuilder(builder)
@@ -87,15 +86,24 @@ func main() {
 		builders[i] = b
 	}
 
-	// set up work environment
-	if err := os.RemoveAll(*buildroot); err != nil {
-		log.Fatalf("Error removing build root (%s): %s", *buildroot, err)
+	if *failAll {
+		failMode(builders)
+		return
 	}
-	if err := os.Mkdir(*buildroot, mkdirPerm); err != nil {
-		log.Fatalf("Error making build root (%s): %s", *buildroot, err)
-	}
-	if err := hgClone(hgUrl, goroot); err != nil {
-		log.Fatal("Error cloning repository:", err)
+
+	// set up work environment, use existing enviroment if possible
+	if hgRepoExists(goroot) {
+		log.Print("Found old workspace, will use it")
+	} else {
+		if err := os.RemoveAll(*buildroot); err != nil {
+			log.Fatalf("Error removing build root (%s): %s", *buildroot, err)
+		}
+		if err := os.Mkdir(*buildroot, mkdirPerm); err != nil {
+			log.Fatalf("Error making build root (%s): %s", *buildroot, err)
+		}
+		if err := hgClone(hgUrl, goroot); err != nil {
+			log.Fatal("Error cloning repository:", err)
+		}
 	}
 
 	if *commitFlag {
@@ -118,14 +126,6 @@ func main() {
 			}
 		}
 		return
-	}
-
-	// external package build mode
-	if *external {
-		if len(builders) != 1 {
-			log.Fatal("only one goos-goarch should be specified with -external")
-		}
-		builders[0].buildExternal()
 	}
 
 	// go continuous build mode (default)
@@ -160,6 +160,21 @@ func main() {
 	}
 }
 
+// go continuous fail mode
+// check for new commits and FAIL them
+func failMode(builders []*Builder) {
+	for {
+		built := false
+		for _, b := range builders {
+			built = b.failBuild() || built
+		}
+		// stop if there was nothing to fail
+		if !built {
+			break
+		}
+	}
+}
+
 func NewBuilder(builder string) (*Builder, error) {
 	b := &Builder{name: builder}
 
@@ -172,7 +187,13 @@ func NewBuilder(builder string) (*Builder, error) {
 	}
 
 	// read keys from keyfile
-	fn := path.Join(os.Getenv("HOME"), ".gobuildkey")
+	fn := ""
+	if runtime.GOOS == "windows" {
+		fn = os.Getenv("HOMEDRIVE") + os.Getenv("HOMEPATH")
+	} else {
+		fn = os.Getenv("HOME")
+	}
+	fn = filepath.Join(fn, ".gobuildkey")
 	if s := fn + "-" + b.name; isFile(s) { // builder-specific file
 		fn = s
 	}
@@ -180,63 +201,14 @@ func NewBuilder(builder string) (*Builder, error) {
 	if err != nil {
 		return nil, fmt.Errorf("readKeys %s (%s): %s", b.name, fn, err)
 	}
-	v := strings.Split(string(c), "\n")
-	b.key = v[0]
-	if len(v) >= 3 {
-		b.codeUsername, b.codePassword = v[1], v[2]
-	}
-
+	b.key = string(bytes.TrimSpace(bytes.SplitN(c, []byte("\n"), 2)[0]))
 	return b, nil
-}
-
-// buildExternal downloads and builds external packages, and
-// reports their build status to the dashboard.
-// It will re-build all packages after pkgBuildInterval nanoseconds or
-// a new release tag is found.
-func (b *Builder) buildExternal() {
-	var prevTag string
-	var nextBuild time.Time
-	for {
-		time.Sleep(waitInterval)
-		err := run(nil, goroot, "hg", "pull", "-u")
-		if err != nil {
-			log.Println("hg pull failed:", err)
-			continue
-		}
-		hash, tag, err := firstTag(releaseRe)
-		if err != nil {
-			log.Println(err)
-			continue
-		}
-		if *verbose {
-			log.Println("latest release:", tag)
-		}
-		// don't rebuild if there's no new release
-		// and it's been less than pkgBuildInterval
-		// nanoseconds since the last build.
-		if tag == prevTag && time.Now().Before(nextBuild) {
-			continue
-		}
-		// build will also build the packages
-		if err := b.buildHash(hash); err != nil {
-			log.Println(err)
-			continue
-		}
-		prevTag = tag
-		nextBuild = time.Now().Add(pkgBuildInterval)
-	}
 }
 
 // build checks for a new commit for this builder
 // and builds it if one is found. 
 // It returns true if a build was attempted.
 func (b *Builder) build() bool {
-	defer func() {
-		err := recover()
-		if err != nil {
-			log.Println(b.name, "build:", err)
-		}
-	}()
 	hash, err := b.todo("build-go-commit", "", "")
 	if err != nil {
 		log.Println(err)
@@ -246,7 +218,6 @@ func (b *Builder) build() bool {
 		return false
 	}
 	// Look for hash locally before running hg pull.
-
 	if _, err := fullHash(goroot, hash[:12]); err != nil {
 		// Don't have hash, so run hg pull.
 		if err := run(nil, goroot, "hg", "pull"); err != nil {
@@ -261,52 +232,39 @@ func (b *Builder) build() bool {
 	return true
 }
 
-func (b *Builder) buildHash(hash string) (err error) {
-	defer func() {
-		if err != nil {
-			err = fmt.Errorf("%s build: %s: %s", b.name, hash, err)
-		}
-	}()
-
+func (b *Builder) buildHash(hash string) error {
 	log.Println(b.name, "building", hash)
 
 	// create place in which to do work
-	workpath := path.Join(*buildroot, b.name+"-"+hash[:12])
-	err = os.Mkdir(workpath, mkdirPerm)
-	if err != nil {
-		return
+	workpath := filepath.Join(*buildroot, b.name+"-"+hash[:12])
+	if err := os.Mkdir(workpath, mkdirPerm); err != nil {
+		return err
 	}
 	defer os.RemoveAll(workpath)
 
 	// clone repo
-	err = run(nil, workpath, "hg", "clone", goroot, "go")
-	if err != nil {
-		return
+	if err := run(nil, workpath, "hg", "clone", goroot, "go"); err != nil {
+		return err
 	}
 
 	// update to specified revision
-	err = run(nil, path.Join(workpath, "go"), "hg", "update", hash)
-	if err != nil {
-		return
+	if err := run(nil, filepath.Join(workpath, "go"), "hg", "update", hash); err != nil {
+		return err
 	}
 
-	srcDir := path.Join(workpath, "go", "src")
+	srcDir := filepath.Join(workpath, "go", "src")
 
 	// build
-	logfile := path.Join(workpath, "build.log")
+	logfile := filepath.Join(workpath, "build.log")
+	cmd := *buildCmd
+	if !filepath.IsAbs(cmd) {
+		cmd = filepath.Join(srcDir, cmd)
+	}
 	startTime := time.Now()
-	buildLog, status, err := runLog(b.envv(), logfile, srcDir, *buildCmd)
+	buildLog, status, err := runLog(b.envv(), logfile, srcDir, cmd)
 	runTime := time.Now().Sub(startTime)
 	if err != nil {
 		return fmt.Errorf("%s: %s", *buildCmd, err)
-	}
-
-	// if we're in external mode, build all packages and return
-	if *external {
-		if status != 0 {
-			return errors.New("go build failed")
-		}
-		return b.buildExternalPackages(workpath, hash)
 	}
 
 	if status != 0 {
@@ -319,83 +277,94 @@ func (b *Builder) buildHash(hash string) (err error) {
 		return fmt.Errorf("recordResult: %s", err)
 	}
 
-	// build goinstallable packages
-	b.buildPackages(filepath.Join(workpath, "go"), hash)
+	// build Go sub-repositories
+	b.buildSubrepos(filepath.Join(workpath, "go"), hash)
 
-	// finish here if codeUsername and codePassword aren't set
-	if b.codeUsername == "" || b.codePassword == "" || !*buildRelease {
-		return
-	}
-
-	// if this is a release, create tgz and upload to google code
-	releaseHash, release, err := firstTag(binaryTagRe)
-	if hash == releaseHash {
-		// clean out build state
-		err = run(b.envv(), srcDir, "./clean.bash", "--nopkg")
-		if err != nil {
-			return fmt.Errorf("clean.bash: %s", err)
-		}
-		// upload binary release
-		fn := fmt.Sprintf("go.%s.%s-%s.tar.gz", release, b.goos, b.goarch)
-		err = run(nil, workpath, "tar", "czf", fn, "go")
-		if err != nil {
-			return fmt.Errorf("tar: %s", err)
-		}
-		err = run(nil, workpath, path.Join(goroot, codePyScript),
-			"-s", release,
-			"-p", codeProject,
-			"-u", b.codeUsername,
-			"-w", b.codePassword,
-			"-l", fmt.Sprintf("%s,%s", b.goos, b.goarch),
-			fn)
-		if err != nil {
-			return fmt.Errorf("%s: %s", codePyScript, err)
-		}
-	}
-
-	return
+	return nil
 }
 
-func (b *Builder) buildPackages(goRoot, goHash string) {
-	for _, pkg := range dashboardPackages() {
+// failBuild checks for a new commit for this builder
+// and fails it if one is found. 
+// It returns true if a build was "attempted".
+func (b *Builder) failBuild() bool {
+	hash, err := b.todo("build-go-commit", "", "")
+	if err != nil {
+		log.Println(err)
+		return false
+	}
+	if hash == "" {
+		return false
+	}
+
+	log.Printf("fail %s %s\n", b.name, hash)
+
+	if err := b.recordResult(false, "", hash, "", "auto-fail mode run by "+os.Getenv("USER"), 0); err != nil {
+		log.Print(err)
+	}
+	return true
+}
+
+func (b *Builder) buildSubrepos(goRoot, goHash string) {
+	for _, pkg := range dashboardPackages("subrepo") {
 		// get the latest todo for this package
 		hash, err := b.todo("build-package", pkg, goHash)
 		if err != nil {
-			log.Printf("buildPackages %s: %v", pkg, err)
+			log.Printf("buildSubrepos %s: %v", pkg, err)
 			continue
 		}
 		if hash == "" {
 			continue
 		}
 
-		// goinstall the package
+		// build the package
 		if *verbose {
-			log.Printf("buildPackages %s: installing %q", pkg, hash)
+			log.Printf("buildSubrepos %s: building %q", pkg, hash)
 		}
-		buildLog, err := b.goinstall(goRoot, pkg, hash)
-		ok := buildLog == ""
+		buildLog, err := b.buildSubrepo(goRoot, pkg, hash)
 		if err != nil {
-			ok = false
-			log.Printf("buildPackages %s: %v", pkg, err)
+			if buildLog == "" {
+				buildLog = err.Error()
+			}
+			log.Printf("buildSubrepos %s: %v", pkg, err)
 		}
 
 		// record the result
-		err = b.recordResult(ok, pkg, hash, goHash, buildLog, 0)
+		err = b.recordResult(err == nil, pkg, hash, goHash, buildLog, 0)
 		if err != nil {
-			log.Printf("buildPackages %s: %v", pkg, err)
+			log.Printf("buildSubrepos %s: %v", pkg, err)
 		}
 	}
 }
 
-func (b *Builder) goinstall(goRoot, pkg, hash string) (string, error) {
-	bin := filepath.Join(goRoot, "bin/goinstall")
+// buildSubrepo fetches the given package, updates it to the specified hash,
+// and runs 'go test -short pkg/...'. It returns the build log and any error.
+func (b *Builder) buildSubrepo(goRoot, pkg, hash string) (string, error) {
+	goBin := filepath.Join(goRoot, "bin")
+	goTool := filepath.Join(goBin, "go")
 	env := append(b.envv(), "GOROOT="+goRoot)
 
+	// add goBin to PATH
+	for i, e := range env {
+		const p = "PATH="
+		if !strings.HasPrefix(e, p) {
+			continue
+		}
+		env[i] = p + goBin + string(os.PathListSeparator) + e[len(p):]
+	}
+
 	// fetch package and dependencies
-	log, status, err := runLog(env, "", goRoot, bin,
-		"-dashboard=false", "-install=false", pkg)
-	if err != nil || status != 0 {
-		return log, err
+	log, status, err := runLog(env, "", goRoot, goTool, "get", "-d", pkg)
+	if err == nil && status != 0 {
+		err = fmt.Errorf("go exited with status %d", status)
+	}
+	if err != nil {
+		// 'go get -d' will fail for a subrepo because its top-level
+		// directory does not contain a go package. No matter, just
+		// check whether an hg directory exists and proceed.
+		hgDir := filepath.Join(goRoot, "src/pkg", pkg, ".hg")
+		if fi, e := os.Stat(hgDir); e != nil || !fi.IsDir() {
+			return log, err
+		}
 	}
 
 	// hg update to the specified hash
@@ -404,8 +373,11 @@ func (b *Builder) goinstall(goRoot, pkg, hash string) (string, error) {
 		return "", err
 	}
 
-	// build the package
-	log, _, err = runLog(env, "", goRoot, bin, "-dashboard=false", pkg)
+	// test the package
+	log, status, err = runLog(env, "", goRoot, goTool, "test", "-short", pkg+"/...")
+	if err == nil && status != 0 {
+		err = fmt.Errorf("go exited with status %d", status)
+	}
 	return log, err
 }
 
@@ -420,8 +392,7 @@ func (b *Builder) envv() []string {
 		"GOROOT_FINAL=/usr/local/go",
 	}
 	for _, k := range extraEnv {
-		s, err := os.Getenverror(k)
-		if err == nil {
+		if s, ok := getenvOk(k); ok {
 			e = append(e, k+"="+s)
 		}
 	}
@@ -433,13 +404,11 @@ func (b *Builder) envvWindows() []string {
 	start := map[string]string{
 		"GOOS":         b.goos,
 		"GOARCH":       b.goarch,
-		"GOROOT_FINAL": "/c/go",
-		// TODO(brainman): remove once we find make that does not hang.
-		"MAKEFLAGS": "-j1",
+		"GOROOT_FINAL": `c:\go`,
+		"GOBUILDEXIT":  "1", // exit all.bat with completion status.
 	}
 	for _, name := range extraEnv {
-		s, err := os.Getenverror(name)
-		if err == nil {
+		if s, ok := getenvOk(name); ok {
 			start[name] = s
 		}
 	}
@@ -492,8 +461,10 @@ func commitWatcher() {
 		if *verbose {
 			log.Printf("poll...")
 		}
+		// Main Go repository.
 		commitPoll(key, "")
-		for _, pkg := range dashboardPackages() {
+		// Go sub-repositories.
+		for _, pkg := range dashboardPackages("subrepo") {
 			commitPoll(key, pkg)
 		}
 		if *verbose {
@@ -534,29 +505,22 @@ var logByHash = map[string]*HgLog{}
 // xmlLogTemplate is a template to pass to Mercurial to make
 // hg log print the log in valid XML for parsing with xml.Unmarshal.
 const xmlLogTemplate = `
-	<log>
-	<hash>{node|escape}</hash>
-	<parent>{parent|escape}</parent>
-	<author>{author|escape}</author>
-	<date>{date|rfc3339date}</date>
-	<desc>{desc|escape}</desc>
-	</log>
+	<Log>
+	<Hash>{node|escape}</Hash>
+	<Parent>{parent|escape}</Parent>
+	<Author>{author|escape}</Author>
+	<Date>{date|rfc3339date}</Date>
+	<Desc>{desc|escape}</Desc>
+	</Log>
 `
 
 // commitPoll pulls any new revisions from the hg server
 // and tells the server about them.
 func commitPoll(key, pkg string) {
-	// Catch unexpected panics.
-	defer func() {
-		if err := recover(); err != nil {
-			log.Printf("commitPoll panic: %s", err)
-		}
-	}()
-
 	pkgRoot := goroot
 
 	if pkg != "" {
-		pkgRoot = path.Join(*buildroot, pkg)
+		pkgRoot = filepath.Join(*buildroot, pkg)
 		if !hgRepoExists(pkgRoot) {
 			if err := hgClone(repoURL(pkg), pkgRoot); err != nil {
 				log.Printf("%s: hg clone failed: %v", pkg, err)
@@ -588,7 +552,7 @@ func commitPoll(key, pkg string) {
 	var logStruct struct {
 		Log []HgLog
 	}
-	err = xml.Unmarshal(strings.NewReader("<top>"+data+"</top>"), &logStruct)
+	err = xml.Unmarshal([]byte("<Top>"+data+"</Top>"), &logStruct)
 	if err != nil {
 		log.Printf("unmarshal hg log: %v", err)
 		return
@@ -662,12 +626,7 @@ func addCommit(pkg, hash, key string) bool {
 }
 
 // fullHash returns the full hash for the given Mercurial revision.
-func fullHash(root, rev string) (hash string, err error) {
-	defer func() {
-		if err != nil {
-			err = fmt.Errorf("fullHash: %s: %s", rev, err)
-		}
-	}()
+func fullHash(root, rev string) (string, error) {
 	s, _, err := runLog(nil, "", root,
 		"hg", "log",
 		"--encoding=utf-8",
@@ -676,7 +635,7 @@ func fullHash(root, rev string) (hash string, err error) {
 		"--template={node}",
 	)
 	if err != nil {
-		return
+		return "", nil
 	}
 	s = strings.TrimSpace(s)
 	if s == "" {
@@ -686,31 +645,6 @@ func fullHash(root, rev string) (hash string, err error) {
 		return "", fmt.Errorf("hg returned invalid hash " + s)
 	}
 	return s, nil
-}
-
-var revisionRe = regexp.MustCompile(`^([^ ]+) +[0-9]+:([0-9a-f]+)$`)
-
-// firstTag returns the hash and tag of the most recent tag matching re.
-func firstTag(re *regexp.Regexp) (hash string, tag string, err error) {
-	o, _, err := runLog(nil, "", goroot, "hg", "tags")
-	for _, l := range strings.Split(o, "\n") {
-		if l == "" {
-			continue
-		}
-		s := revisionRe.FindStringSubmatch(l)
-		if s == nil {
-			err = errors.New("couldn't find revision number")
-			return
-		}
-		if !re.MatchString(s[1]) {
-			continue
-		}
-		tag = s[1]
-		hash, err = fullHash(goroot, s[2])
-		return
-	}
-	err = errors.New("no matching tag found")
-	return
 }
 
 var repoRe = regexp.MustCompile(`^code\.google\.com/p/([a-z0-9\-]+(\.[a-z0-9\-]+)?)(/[a-z0-9A-Z_.\-/]+)?$`)
@@ -723,4 +657,27 @@ func repoURL(importPath string) string {
 		return ""
 	}
 	return "https://code.google.com/p/" + m[1]
+}
+
+// defaultSuffix returns file extension used for command files in
+// current os environment.
+func defaultSuffix() string {
+	if runtime.GOOS == "windows" {
+		return ".bat"
+	}
+	return ".bash"
+}
+
+func getenvOk(k string) (v string, ok bool) {
+	v = os.Getenv(k)
+	if v != "" {
+		return v, true
+	}
+	keq := k + "="
+	for _, kv := range os.Environ() {
+		if kv == keq {
+			return "", true
+		}
+	}
+	return "", false
 }
